Move AWS create options to separate files
